+++ IOStream/FileInputStream.m
@@ -21,7 +21,7 @@
 
 @implementation FileInputStream
 
-- (id)initWithFilename:(NSString*)filename {
+- (id)initWithFilename:(NSString *)filename {
     self = [super init];
     if (self) {
         fd = open([filename UTF8String], O_RDONLY);
@@ -36,11 +36,11 @@
     [self close];
 }
 
-- (NSUInteger)read:(void*)bytes length:(NSUInteger)bytesLength {
+- (NSUInteger)read:(void *)bytes length:(NSUInteger)bytesLength {
     return read(fd, bytes, bytesLength);
 }
 
-- (NSInteger)seek:(NSUInteger)offset {
+- (off_t)seek:(off_t)offset {
     return lseek(fd, offset, SEEK_SET);
 }
